requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
discord.py==1.5.0
docutils==0.16
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
keyring==21.2.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
multidict==4.7.6
packaging==20.4
Pillow==7.2.0
pkginfo==1.5.0.1
pluggy==0.13.1
py==1.9.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
python-dotenv==0.14.0
pywin32-ctypes==0.2.0
readme-renderer==26.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
toml==0.10.1
tqdm==4.48.0
twine==3.2.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.1.0
